Moved Section components outside of "containers" folder, as it makes more sense
