+++ src/components/MainInformationSection/AboutSection/AboutSection.styled.tsx
