+++ src/components/MainInformationSection/MainInformationSection.styled.tsx
